Update settings.xml
Removed jboss repo to be consistent with testing repositories for product. To test if we have provided the dependency in product repository. Customers should be able to use Central and private MRRC.
Kogito group are at /kogito-master-nightly/ it contains
Central + JBoss.org public repo + scratch-release-kogito-master

It is recommended to avoid adding JBoss.org public repo in settings directly since it contains project SNAPSHOTS and we would not be able to catch issues if some snapshot was not updated at the product pipeline.

Use https fpr Maven central