Add native for quickstarts and optaplanner quarkus integration
Quickstarts are collected in different places of the artifact repository since the pojects are completely undependent I decide to get rid of filtering. Also I added non-native so to not append pomincludes together in case of poms located in different folders

Optaplanner profiles is just the same as in kogito.